+++ docs/src/js/docs.js
@@ -9,6 +9,9 @@ shr.setup({
 	debug: 	true,
 	count: {
 		//position: 'before'
+	},
+	tokens: {
+		github: 'a46ac653210ba6a6be44260c29c333470c3fbbf5'
 	}
 });