+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.1.5 - 0.1.6
+- Fixes for IE9
+- Fixed toLocaleString() for old IE
+- Added local storage caching
+- Added support for GitHub (stars, forks, etc)
+- Jazzed up docs
+
 ## v0.1.4
 - Docs improvements